+++ droidbot/utils.py
@@ -1,4 +1,5 @@
 import re
+import signal
 from datetime import datetime
 
 # logcat regex, which will match the log message generated by `adb logcat -v threadtime`
@@ -28,3 +29,27 @@ def parse_log(log_msg):
     log_dict['datetime'] = datetime.strptime(datetime_str, "%Y-%m-%d %H:%M:%S.%f")
 
     return log_dict
+
+
+class Timeout:
+    def __init__(self, seconds=0, error_message='Timeout'):
+        self.seconds = seconds
+        self.error_message = error_message
+
+    def handle_timeout(self, signum, frame):
+        raise TimeoutException()
+
+    def __enter__(self):
+        if self.seconds != 0:
+            signal.signal(signal.SIGALRM, self.handle_timeout)
+            signal.alarm(self.seconds)
+
+    def __exit__(self, type, value, traceback):
+        signal.alarm(0)
+
+
+class TimeoutException(Exception):
+    """
+    Exception if connection has been waiting too long
+    """
+    pass
